test setup moving query up level
