+++ app/javascript/containers/bookmarkr/BookmarkrApp.jsx
@@ -8,20 +8,22 @@ import ConfigProvider from './containers/ConfigProvider';
 //* Hook Imports
 import fetchConfig from './containers/hooks/fetchConfig';
 
-export default function BookmarkrApp() {
-	const { isLoading, error, data } = useQuery('stateConfig', fetchConfig);
+export default function BookmarkrApp(props) {
+	// const { isLoading, error, data } = useQuery('stateConfig', fetchConfig);
 
-	const configFormatter = (data) => {
-		if (!data) return;
-		return Object.entries(data.data.attributes).map((obj) => ({ [obj[0]]: obj[1] }));
-	};
+	// const configFormatter = (data) => {
+	// 	if (!data) return;
+	// 	return Object.entries(data.data.attributes).map((obj) => ({ [obj[0]]: obj[1] }));
+	// };
+	debugger
 
 	return (
 		<>
 			<div className='h-screen flex flex-col'>
-				{isLoading && <>Loading...</>}
+				Booksmark the spot
+				{/* {isLoading && <>Loading...</>}
 				{error && <>Uh oh we done goofed...</>}
-				{!isLoading && <ConfigProvider config={configFormatter(data)} />}
+				{!isLoading && <ConfigProvider config={configFormatter(data)} />} */}
 			</div>
 		</>
 	);
