+++ app/javascript/packs/bookmarkr.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import BookmarkrApp from '../containers/bookmarkr/BookmarkrApp';
-import { QueryClient, QueryClientProvider } from 'react-query';
+import { QueryClient, QueryClientProvider, useQuery } from 'react-query';
 
 const queryClient = new QueryClient({
 	defaultOptions: {
@@ -11,10 +11,18 @@ const queryClient = new QueryClient({
 	},
 });
 
+const configFormatter = (data) => {
+	if (!data) return;
+	return Object.entries(data.data.attributes).map((obj) => ({ [obj[0]]: obj[1] }));
+};
+
+const { isLoading, error, data } = useQuery('stateConfig', fetchConfig);
+
 ReactDOM.render(
 	<React.StrictMode>
 		<QueryClientProvider client={queryClient}>
-			<BookmarkrApp />
+			{isLoading && <>Loading...</>}
+			{!isLoading && <BookmarkrApp config={configFormatter(data)}/>}
 		</QueryClientProvider>
 	</React.StrictMode>,
 	document.getElementById('homeContainer')
